tradingbot/requirements.txt
# Core
py-clob-client>=0.15.0
ccxt>=4.0.0
web3==6.14.0
httpx>=0.27.0
python-dotenv>=1.0.0

# AI
anthropic>=0.40.0

# Math
numpy>=1.26.0
scipy>=1.12.0

# Monitoring
aiogram>=3.0.0
sentry-sdk>=2.0.0
structlog>=24.0.0

# Database
supabase>=2.0.0
aiosqlite>=0.19.0

# API (for dashboard communication)
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
